feat : separate classes for local detection / triton detection
