+++ test-local-inference.py
@@ -9,11 +9,27 @@ from PIL import Image
 import torch
 
     
-from dptr.trbaOcr import TrbaOCR 
+#from dptr.trbaOcr import TrbaOCR 
 
 saved_model = '/home/raki-dedigama/projects/rr/model-training/ean/models/trained-models/best_accuracy.pt'
-trbaOCR = TrbaOCR(device='cuda')    
-trbaOCR.load_model_from_disk(saved_model)
+#trbaOCR = TrbaOCR(device='cuda')    
+#trbaOCR.load_model_from_disk(saved_model)
+
+from dptr.trba_detector import TRBADetector
+trba_detector = TRBADetector()
+trba_detector.load_model_from_disk(saved_model)
+
+# Run from image path
+# pil_image = Image.open(image_path)#   
+# predicted_text = trbaOCR.img_to_ean(pil_image)
+# print("predicted_text : ", ean['pred'], ean['score'])
+
+image_path = 'demo_image/ean.jpg'
+pil_image = Image.open(image_path)
+
+#ean = trbaOCR.img_path_to_ean(image_path)
+ean = trba_detector.predict(pil_image)
+print("predicted : ", ean['pred'], ean['score'])
 
 # ## Trace model to torchscript        
 # model = trbaOCR.model
@@ -32,10 +48,7 @@ trbaOCR.load_model_from_disk(saved_model)
 
 # exit()
 
-# Run from image path
-image_path = 'demo_image/ean.jpg'
-ean = trbaOCR.img_path_to_ean(image_path)
-print("predicted : ", ean['pred'], ean['score'])
+
 #trbaOCR.convert_model_to_onnx()
