+++ triton.json
@@ -1,0 +1,60 @@
+{
+	"ShelfObjects" :
+	{
+		"TRITON_REGION_INFERENCE_SERVER_URL":"localhost:8000",
+		"TRITON_REGION_INFERENCE_PROTOCOL": "HTTP",	
+		"TRITON_REGION_INFERENCE_FLAG_VERBOSE": false,		
+		"TRITON_REGION_INFERENCE_MODEL_NAME": "yolov4_onnx",
+		"TRITON_REGION_INFERENCE_MODEL_VERSION":"",
+		"TRITON_REGION_INFERENCE_BATCH_SIZE":1,
+		"TRITON_REGION_INFERENCE_CLASSES": 1,
+		"TRITON_REGION_INFERENCE_IMAGE_SCALING": null,
+		"TRITON_REGION_INFERENCE_ASYNC_SET": false,
+		"TRITON_REGION_INFERENCE_STREAMING": false
+	
+	},
+	"PriceTagSubRegions" :
+	{
+		"TRITON_REGION_INFERENCE_SERVER_URL":"localhost:8000",
+		"TRITON_REGION_INFERENCE_PROTOCOL": "HTTP",	
+		"TRITON_REGION_INFERENCE_FLAG_VERBOSE": false,		
+		"TRITON_REGION_INFERENCE_MODEL_NAME": "ptsr_onnx",
+		"TRITON_REGION_INFERENCE_MODEL_VERSION":"",
+		"TRITON_REGION_INFERENCE_BATCH_SIZE":1,
+		"TRITON_REGION_INFERENCE_CLASSES": 1,
+		"TRITON_REGION_INFERENCE_IMAGE_SCALING": null,
+		"TRITON_REGION_INFERENCE_ASYNC_SET": false,
+		"TRITON_REGION_INFERENCE_STREAMING": false
+	
+	},
+	"ShelfModuleMarkerSubRegions" :
+	{
+		"TRITON_REGION_INFERENCE_SERVER_URL":"localhost:8000",
+		"TRITON_REGION_INFERENCE_PROTOCOL": "HTTP",	
+		"TRITON_REGION_INFERENCE_FLAG_VERBOSE": false,		
+		"TRITON_REGION_INFERENCE_MODEL_NAME": "smm_onnx",
+		"TRITON_REGION_INFERENCE_MODEL_VERSION":"",
+		"TRITON_REGION_INFERENCE_BATCH_SIZE":1,
+		"TRITON_REGION_INFERENCE_CLASSES": 1,
+		"TRITON_REGION_INFERENCE_IMAGE_SCALING": null,
+		"TRITON_REGION_INFERENCE_ASYNC_SET": false,
+		"TRITON_REGION_INFERENCE_STREAMING": false
+	
+	},
+	"EANs":
+	{
+		"TRITON_OCR_SERVER_URL":"localhost:8000",
+		"TRITON_OCR_PROTOCOL": "HTTP",	
+		"TRITON_OCR_FLAG_VERBOSE": false,		
+		"TRITON_OCR_MODEL_NAME": "ean_pt",
+		"TRITON_OCR_MODEL_VERSION":"",
+		"TRITON_OCR_BATCH_SIZE":1,
+		"TRITON_OCR_CLASSES": 1,
+		"TRITON_OCR_IMAGE_SCALING": null,
+		"TRITON_OCR_ASYNC_SET": false,
+		"TRITON_OCR_STREAMING": false
+	}
+	
+
+
+}
